refactor: 불필요 코드 제거 및 정리
