+++ build.gradle
@@ -19,11 +19,9 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
 
-	//Swagger, RestDocs
+	//Swagger
 	implementation group: 'org.springdoc', name: 'springdoc-openapi-starter-webmvc-ui', version: '2.1.0'
 	testImplementation group: 'org.springdoc', name: 'springdoc-openapi-starter-webmvc-api', version: '2.1.0'
-
-	implementation 'org.springframework.cloud:spring-cloud-starter-openfeign:4.1.0'
 
 	//S3
 	implementation 'org.springframework.cloud:spring-cloud-starter-aws:2.2.6.RELEASE'
@@ -41,7 +39,6 @@ dependencies {
 
 	// http client
 	implementation 'org.apache.httpcomponents.client5:httpclient5:5.3.1'
-
 
 	//db
 	implementation group: 'mysql', name: 'mysql-connector-java', version: '8.0.32'
@@ -63,7 +60,6 @@ dependencies {
 	//Monitoring
 	implementation 'org.springframework.boot:spring-boot-starter-actuator'
 	implementation 'io.micrometer:micrometer-registry-prometheus'
-
 }
 
 tasks.named('test') {
