+++ src/main/java/com/shallwe/domain/auth/application/AuthService.java
@@ -5,11 +5,12 @@ import java.util.Optional;
 import com.shallwe.domain.auth.domain.AppleToken;
 import com.shallwe.domain.auth.domain.repository.AppleTokenRepository;
 import com.shallwe.domain.auth.dto.*;
+import com.shallwe.domain.auth.dto.request.*;
+import com.shallwe.domain.auth.dto.response.AuthRes;
 import com.shallwe.domain.auth.exception.*;
 import com.shallwe.domain.common.Status;
 import com.shallwe.domain.shopowner.domain.ShopOwner;
 import com.shallwe.domain.shopowner.domain.repository.ShopOwnerRepository;
-import com.shallwe.domain.auth.dto.ShopOwnerChangePasswordReq;
 import com.shallwe.domain.shopowner.exception.AlreadyExistPhoneNumberException;
 import com.shallwe.domain.shopowner.exception.InvalidPhoneNumberException;
 import com.shallwe.domain.shopowner.exception.InvalidShopOwnerException;
