+++ src/main/java/com/shallwe/domain/auth/dto/request/AppleTokenReq.java
@@ -1,4 +1,4 @@
-package com.shallwe.domain.auth.dto;
+package com.shallwe.domain.auth.dto.request;
 
 import lombok.Data;
