+++ src/main/java/com/shallwe/domain/auth/dto/request/ShopOwnerChangePasswordReq.java
@@ -1,4 +1,4 @@
-package com.shallwe.domain.auth.dto;
+package com.shallwe.domain.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
