+++ src/main/java/com/shallwe/domain/auth/dto/response/AuthRes.java
@@ -1,4 +1,4 @@
-package com.shallwe.domain.auth.dto;
+package com.shallwe.domain.auth.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
